create domain and presentation packages
