+++ lib/main.dart
@@ -1,10 +1,6 @@
-import 'package:check_palindrome/presentation/app.dart';
 import 'package:flutter/material.dart';
+import 'package:presentation/app.dart';
 
 void main() {
   runApp(const MyApp());
 }
-
-
-
-
