--- packages/check_palindrome/analysis_options.yaml
@@ -1,28 +1,0 @@
-include: package:flutter_lints/flutter.yaml
-
-# Additional information about this file can be found at
-# https://dart.dev/guides/language/analysis-options
-
-linter:
-
-  rules:
-    - avoid_types_on_closure_parameters
-    - avoid_void_async
-    - camel_case_types
-    - cancel_subscriptions
-    - close_sinks
-    - constant_identifier_names
-    - control_flow_in_finally
-    - directives_ordering
-    - empty_statements
-    - hash_and_equals
-    - implementation_imports
-    - non_constant_identifier_names
-    - package_api_docs
-    - package_names
-    - package_prefixed_library_names
-    - prefer_const_constructors
-    - prefer_const_literals_to_create_immutables
-    - test_types_in_equals
-    - throw_in_finally
-    - unnecessary_statements
