--- packages/check_palindrome/lib/data/text_data.dart
@@ -1,4 +1,0 @@
-class DataPalindrome {
-  String text = '';
-  bool answer = false;
-}
