+++ packages/domain/lib/use_case/use_case.dart
@@ -1,0 +1,3 @@
+abstract class UseCase {
+  bool call();
+}
