+++ packages/presentation/pubspec.yaml
@@ -10,6 +10,8 @@ environment:
 dependencies:
   flutter:
     sdk: flutter
+  domain:
+    path: /Users/kabak/StudioProjects/palindrome/packages/domain
 
 dev_dependencies:
   flutter_test:
